[scheduler] Move the Event Timeline slot types to the premium package
`EventTimelineSlots`, `EventTimelineSlotProps` and the slot props
interfaces are exported from `@mui/x-scheduler-premium/models`, next
to the component they describe. The community package keeps the shared
`SchedulerSlots`.
