+++ docs/data/scheduler/event-timeline/events/EventContentSlot.tsx
@@ -1,12 +1,9 @@
 import * as React from 'react';
 import FlagRounded from '@mui/icons-material/FlagRounded';
 import BuildRounded from '@mui/icons-material/BuildRounded';
-import {
-  SchedulerEvent,
-  SchedulerResource,
-  TimelineEventContentProps,
-} from '@mui/x-scheduler/models';
+import { SchedulerEvent, SchedulerResource } from '@mui/x-scheduler/models';
 import { EventTimelinePremium } from '@mui/x-scheduler-premium/event-timeline-premium';
+import { TimelineEventContentProps } from '@mui/x-scheduler-premium/models';
 import { defaultVisibleDate } from '../../datasets/company-roadmap';
 
 interface DemoEvent extends SchedulerEvent {
