+++ docs/data/scheduler/event-timeline/resources/ResourceTitleSlot.js
@@ -3,6 +3,7 @@ import Tooltip from '@mui/material/Tooltip';
 import Link from '@mui/material/Link';
 
 import { EventTimelinePremium } from '@mui/x-scheduler-premium/event-timeline-premium';
+
 import { defaultVisibleDate } from '../../datasets/company-roadmap';
 
 const resourceDetails = {
